Revert "Apparently we can't support size_t's z modifier in formats."
This reverts (most of) commit ef3b5fe24263f47db8ef937f450cd3655c1b59b4.

BIG FAT NOTE: We provide a C99 snprintf(). So our Log() function and
              most other libutils facilies *are perfectly good* with C99
              modifiers. You should only avoid C99 modifiers when
              calling non-snprintf functions directly from libc, e.g. an
              fprintf() call.

Conflicts:
	cf-serverd/tls_server.c
	libpromises/dbm_tokyocab.c
	libpromises/eval_context.c
	libpromises/evalfunction.c
	libpromises/policy.c
	libpromises/promises.c

(cherry picked from commit dc36265712c0360d5b73db99e8d3def861b22a50)

+++ cf-agent/tokyo_check.c
@@ -187,8 +187,8 @@ static int DBMetaPopulateOffsetMap(DBMeta * dbmeta)
         }
     }
 
-    Log(LOG_LEVEL_VERBOSE, "Found %llu buckets with offsets",
-        (unsigned long long) StringMapSize(dbmeta->offset_map));
+    Log(LOG_LEVEL_VERBOSE, "Found %zu buckets with offsets",
+            StringMapSize(dbmeta->offset_map));
     return 0;
 }
